refactor(core): lift the track-viewing commands into mxb-core
Stage 6a — the low-risk half of the viewer lift, done on its own so the model
loaders land against a boundary that is already known to work.

Thirteen commands move to `mxb_core::trackview`: the `.pkz` metadata readers and
everything that answers "what does this track look like" — its info block, its
terrain, its scenery, the sheets its ground is painted with. None of them writes
anything, and both binaries need them: the manager shows a track in the Library,
the studio previews one it is building.

This is also the first proof that cross-crate command registration works.
`generate_handler!` builds the path to a macro the `#[tauri::command]` expansion
re-exports at the function's own path with the function's own visibility, so the
commands are `pub` and the handler names them as
`mxb_core::trackview::read_track_info`. The codebase already did this for
`plugins::*`; this is the same mechanism across a crate boundary.

Adds scripts/command-census.sh, run in CI. The compiler cannot check the handler
list: it takes tokens, so a name dropped from it is not a build error but a
runtime "command not found" on whichever screen needs it. Moving commands between
crates is exactly what invites that, so it gets its own check.

The census found real dead code on its first run: `frostmod_keybinds`,
`frostmod_set_keybinds` and `frostmod_default_keybinds` are called from
`api/mods.ts` but are defined nowhere in Rust, and nothing calls those wrappers
either. A feature removed on the Rust side that left its API surface behind.
Removed, along with the `FrostmodKeybind` type nothing used.

Two things about the census itself are worth knowing, because both were silent:
BSD sed has no `\s`, so the first version's handler range ran to end of file and
swept up struct fields; and requiring a trailing comma drops whichever command is
last in the list, which is how `mods_state_restore_all` came out "unregistered".
It is node now, and the two directions differ in strictness on purpose — only a
literal inside `invoke(...)` can fail the build, because a false positive there
is a broken build for nothing, while the informational list also counts names
that reach `invoke` indirectly (`loadTrackSurfaces` takes its command as a
defaulted parameter).

Verified: cargo check clean on host and x86_64-pc-windows-gnu; 1178 + 309 tests
pass, unchanged; census green; lint clean.
